🔨 Refactor cyclonedx (#9668)
* add json and xml

* fix xml part

* fix json part

* add helper

* advance helper

* fix_severity into helpers

* update

* flake8

* remove internal_deduplicate as it is unused
